#11 디렉토리 구조 변경
디렉토리 구조를 도메인 중심으로 변경하여 기존 파일들의 경로를 수정하였다.
