+++ src/main/java/dev/sodev/LoggingController.java
@@ -1,0 +1,23 @@
+package dev.sodev;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class LoggingController {
+
+    @Value("${logging-module.version}")
+    private String version;
+
+    @GetMapping("/v1")
+    public String version() {
+        return String.format("Project Version : %s", version);
+    }
+
+    @GetMapping("/health")
+    public String checkHealth() {
+        return "healthy";
+    }
+
+}
