--- src/main/java/dev/sodev/domain/entity/AlarmArgs.java
@@ -1,13 +1,0 @@
-package dev.be.sodevcommon.model.entity;
-
-
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-public class AlarmArgs {
-
-    private Long fromUserId;
-    private Long targetId;
-}
