--- src/main/java/dev/sodev/domain/entity/Member.java
@@ -1,80 +1,0 @@
-package dev.sodev.domain.entity;
-
-import dev.sodev.domain.enums.Auth;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@SQLDelete(sql = "UPDATE \"member\" SET removed_at = NOW() WHERE id=?")
-@Where(clause = "removed_at is NULL")
-@Entity
-public class Member extends BaseEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_id")
-    private Long id;
-    private String email;
-    private String password;
-
-    @Builder.Default
-    @Enumerated(EnumType.STRING)
-    private Auth auth = Auth.MEMBER;
-
-    private String nickName;
-    private String phone;
-    private String introduce;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_project_id")
-    private MemberProject memberProject;
-
-    private String imageUrl;
-    private String imageUuid;
-    private String imageName;
-
-    @Column(length = 1000)
-    private String refreshToken;
-
-
-    private LocalDateTime removedAt;
-
-    //== 정보 수정 ==//
-    public void updatePassword(PasswordEncoder passwordEncoder, String password){
-        this.password = passwordEncoder.encode(password);
-    }
-
-    public void updateName(String name){
-        this.email = name;
-    }
-
-    public void updateNickName(String nickName){
-        this.nickName = nickName;
-    }
-
-    public void updateRefreshToken(String refreshToken){
-        this.refreshToken = refreshToken;
-    }
-
-    public void destroyRefreshToken(){
-        this.refreshToken = null;
-    }
-
-
-
-    //== 패스워드 암호화 ==//
-    public void encodePassword(PasswordEncoder passwordEncoder){
-        this.password = passwordEncoder.encode(password);
-    }
-
-}
