--- src/main/java/dev/sodev/domain/entity/MemberProject.java
@@ -1,29 +1,0 @@
-package dev.sodev.domain.entity;
-
-
-import dev.sodev.domain.enums.ProjectState;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class MemberProject extends BaseEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "member_project_id")
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "project_id")
-    private Project project;
-
-    @Enumerated(EnumType.STRING)
-    private ProjectState state;
-
-}
