--- src/main/java/dev/sodev/domain/entity/MessageImage.java
@@ -1,11 +1,0 @@
-package dev.sodev.domain.entity;
-
-import jakarta.persistence.Entity;
-
-@Entity
-public class MessageImage extends Message {
-
-    private String url;
-    private String name;
-    private String uuid;
-}
