--- src/main/java/dev/sodev/domain/entity/MessageText.java
@@ -1,9 +1,0 @@
-package dev.sodev.domain.entity;
-
-import jakarta.persistence.Entity;
-
-@Entity
-public class MessageText extends Message {
-
-    private String text;
-}
