--- src/main/java/dev/sodev/domain/entity/Project.java
@@ -1,31 +1,0 @@
-package dev.sodev.domain.entity;
-
-import dev.sodev.domain.enums.ProjectState;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-import java.time.LocalDateTime;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class Project extends BaseEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "project_id")
-    private Long id;
-    private String title;
-    private String content;
-    private Integer be;
-    private Integer fe;
-
-    @Enumerated(EnumType.STRING)
-    private ProjectState state;
-
-    private LocalDateTime startDate;
-    private LocalDateTime endDate;
-    private LocalDateTime recruitDate;
-
-
-}
