--- src/main/java/dev/sodev/domain/entity/Skill.java
@@ -1,20 +1,0 @@
-package dev.sodev.domain.entity;
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-public class Skill extends BaseEntity{
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "skill_id")
-    private Long id;
-
-    private String name;
-
-    private Long usage;
-
-}
