+++ src/main/java/dev/sodev/domain/member/service/LoginService.java
@@ -1,0 +1,31 @@
+package dev.sodev.domain.member.service;
+
+import dev.sodev.domain.member.Member;
+import dev.sodev.global.exception.ErrorCode;
+import dev.sodev.global.exception.SodevApplicationException;
+import dev.sodev.domain.member.repository.MemberRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service
+@RequiredArgsConstructor
+public class LoginService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String memberEmail) throws UsernameNotFoundException {
+
+        Member member = memberRepository.findByEmail(memberEmail).orElseThrow(() -> new SodevApplicationException(ErrorCode.EMAIL_NOT_FOUND, "존재하지 않는 이메일 입니다."));
+
+        return User.builder().username(member.getEmail())
+                .password(member.getPassword())
+                .roles(member.getAuth().name())
+                .build();
+    }
+
+}
