+++ src/main/java/dev/sodev/domain/message/MessageImage.java
@@ -1,0 +1,14 @@
+package dev.sodev.domain.message;
+
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MessageImage extends Message {
+
+    private String url;
+    private String name;
+    private String uuid;
+}
