+++ src/main/java/dev/sodev/domain/message/MessageText.java
@@ -1,0 +1,12 @@
+package dev.sodev.domain.message;
+
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class MessageText extends Message {
+
+    private String text;
+}
