+++ src/main/java/dev/sodev/domain/project/Project.java
@@ -1,0 +1,33 @@
+package dev.sodev.domain.project;
+
+import dev.sodev.domain.BaseEntity;
+import dev.sodev.domain.enums.ProjectState;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Project extends BaseEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "project_id")
+    private Long id;
+    private String title;
+    private String content;
+    private Integer be;
+    private Integer fe;
+
+    @Enumerated(EnumType.STRING)
+    private ProjectState state;
+
+    private LocalDateTime startDate;
+    private LocalDateTime endDate;
+    private LocalDateTime recruitDate;
+
+
+}
