+++ src/main/java/dev/sodev/domain/skill/Skill.java
@@ -1,0 +1,22 @@
+package dev.sodev.domain.skill;
+
+import dev.sodev.domain.BaseEntity;
+import jakarta.persistence.*;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class Skill extends BaseEntity {
+
+    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "skill_id")
+    private Long id;
+
+    private String name;
+
+    private Long usage;
+
+}
