+++ src/main/java/dev/sodev/global/Response.java
@@ -1,0 +1,36 @@
+package dev.sodev.global;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class Response<T> {
+    private String resultCode;
+    private T result;
+
+    public static <T> Response<T> success() {
+        return new Response<T>("SUCCESS", null);
+    }
+
+    public static <T> Response<T> success(T result) {
+        return new Response<T>("SUCCESS", result);
+    }
+
+    public static Response<Void> error(String resultCode) {
+        return new Response<Void>(resultCode, null);
+    }
+
+    public String toStream() {
+        if (result == null) {
+            return "{" +
+                    "\"resultCode\":" + "\"" + resultCode + "\"," +
+                    "\"result\":" + null +
+                    "}";
+        }
+        return "{" +
+                "\"resultCode\":" + "\"" + resultCode + "\"," +
+                "\"result\":" + "\"" + result + "\"," +
+                "}";
+    }
+}
