+++ src/main/java/dev/sodev/global/util/SecurityUtil.java
@@ -1,0 +1,11 @@
+package dev.sodev.global.util;
+
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
+
+public class SecurityUtil {
+
+    public static String getMemberEmail() {
+        return ((UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
+    }
+}
