--- src/main/java/dev/sodev/repository/MemberRepository.java
@@ -1,13 +1,0 @@
-package dev.sodev.repository;
-
-import dev.sodev.domain.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-
-    Optional<Member> findByEmail(String email);
-
-    Optional<Member> findByRefreshToken(String refreshToken);
-}
