--- src/main/java/dev/sodev/util/SecurityUtil.java
@@ -1,11 +1,0 @@
-package dev.sodev.util;
-
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.core.userdetails.UserDetails;
-
-public class SecurityUtil {
-
-    public static String getLoginEmail() {
-        return ((UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal()).getUsername();
-    }
-}
